Create an AuthorizedExecutableDelegate protocol
Declare FinkController conforming to the protocol.
Also fix executableFinished:withStatus: to expect an NSNumber not just an int.
This is why protocols are important for implementing delegates; it allows the compiler to catch errors like this. :)

+++ AuthorizedExecutable.m
@@ -10,6 +10,7 @@
 */
 
 #import "AuthorizedExecutable.h"
+#import "FinkController.h"
 #import <Security/AuthorizationTags.h>
 
 @implementation AuthorizedExecutable
@@ -434,7 +435,7 @@
 //
 - (void)stop
 {
-	int status;
+	NSInteger status;
 
     if (stdoutHandle)
     {
